requirements.txt
python-dotenv==1.1.0
requests==2.32.3
terminaltables==3.1.10